+++ src/dfx/src/lib/mod.rs
@@ -1,6 +1,7 @@
 pub mod agent;
 pub mod builders;
 pub mod canister_info;
+pub mod cycles_ledger_types;
 pub mod deps;
 pub mod diagnosis;
 pub mod dist;